src/ehr-fhir/requirements.txt
requests
google-cloud-spanner
google-cloud-kms
